Let clairnote-sn-switch.js modify an item in first/home menu
The `More` menu is now gone and its contents are under the first
menu item (home / `Clairnote Music Notation`).  So we edit the JS
for dynamically modifying the menu to allow switching between
Clairnote and Clairnote SN to fit the new menu structure.

Rather than `Switch to Clairnote SN site` or similar, we use
`Clairnote SN - Current Page` in parallel to `Clairnote - Home`.